+++ freemarker-generator-cli/CHANGELOG.md
@@ -17,6 +17,9 @@ All notable changes to this project will be documented in this file. We try to a
 * [FREEMARKER-129] Migrate `freemarker-cli` into `freemarker-generator` project (see [https://github.com/sgoeschl/freemarker-cli](https://github.com/sgoeschl/freemarker-cli))
 
 ### Changed
+* [FREEMARKER-153] Packaged templates are now prefixed with `freemarker-generator`, e.g. `freemarker-generator/info.ftl`
+* [FREEMARKER-153] Renamed `--basedir` command line option to `--template-dir`
+* [FREEMARKER-153] Renamed `freemarker-cli` to `freemarker-generator`
 * Removing `DataSources.first` and use `dataSources.get(0)` instead
 * [FREEMARKER-146] Cleanly separate example templates and data from user-supplied content
 * `DataSource` use `uri` instead of `location`
@@ -31,6 +34,9 @@ All notable changes to this project will be documented in this file. We try to a
 * [FREEMARKER-151] Ensure that build and and examples are running on Windows
 * [FREEMARKER-147] Complete Maven site documentation
 * [FREEMARKER-127] Site build fails with missing "org/apache/maven/doxia/siterenderer/DocumentContent"
+
+### Internal
+* [FREEMARKER-153] FreeMarker Generator release preparations
 
 [FREEMARKER-127]: https://issues.apache.org/jira/browse/FREEMARKER-127
 [FREEMARKER-128]: https://issues.apache.org/jira/browse/FREEMARKER-128
@@ -47,3 +53,4 @@ All notable changes to this project will be documented in this file. We try to a
 [FREEMARKER-148]: https://issues.apache.org/jira/browse/FREEMARKER-148
 [FREEMARKER-149]: https://issues.apache.org/jira/browse/FREEMARKER-149
 [FREEMARKER-151]: https://issues.apache.org/jira/browse/FREEMARKER-151
+[FREEMARKER-153]: https://issues.apache.org/jira/browse/FREEMARKER-151